requirements.txt
wheel
django
djangorestframework
markdown
django-filter
pytils
Pillow
aiogram
django-admin-vali
django-admin-shortcuts
psycopg2
requests
python-telegram-bot-pagination